modulo ordenes de compras

+++ database/migrations/2025_12_01_195519_create_remito_items_table.php
@@ -9,18 +9,20 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-   public function up()
+    public function up()
     {
-        Schema::create('remito_items', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('remito_id');
-            $table->string('descripcion');
-            $table->integer('cantidad')->default(1);
+        if (!Schema::hasTable('remito_items')) {
+            Schema::create('remito_items', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('remito_id');
+                $table->string('descripcion');
+                $table->integer('cantidad')->default(1);
 
-            $table->timestamps();
+                $table->timestamps();
 
-            $table->foreign('remito_id')->references('id')->on('remitos')->onDelete('cascade');
-        });
+                $table->foreign('remito_id')->references('id')->on('remitos')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
